Added simple API for User and Message entities

+++ spring-data/src/main/java/com/volvadvit/springdata/exception/UserNotFoundException.java
@@ -1,7 +1,7 @@
 package com.volvadvit.springdata.exception;
 
 public class UserNotFoundException extends RuntimeException {
-    public UserNotFoundException(final Long id) {
+    public UserNotFoundException(final int id) {
         super("User with id " + id + " not found");
     }
 }
